secret: rename from secrets to avoid conflict with db.secrets
