+++ internal/secret/encryptor.go
@@ -1,4 +1,4 @@
-package secrets
+package secret
 
 import (
 	"crypto/aes"
