+++ internal/secret/init.go
@@ -1,4 +1,4 @@
-package secrets
+package secret
 
 import (
 	"bytes"
